requirements.txt
dictdiffer==0.6.1
ipython==6.1.0
isort==4.2.15
opbeat==3.5.1
pycodestyle==2.3.1
pydocstyle==2.0.0
pymongo==3.4.0
python-dotenv==0.6.0
requests==2.11.1
